refactor(currencies.service): adicionando validações no crud de currencies

+++ src/currencies/currencies.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateCurrencyDto } from './dto/create-currency.dto';
 import { UpdateCurrencyDto } from './dto/update-currency.dto';
 import { PrismaService } from '../prisma_client/prisma.service';
@@ -8,10 +12,14 @@ export class CurrenciesService {
   constructor(private prisma: PrismaService) {}
 
   async create(data: CreateCurrencyDto): Promise<CreateCurrencyDto> {
+    if (!data.name) {
+      throw new BadRequestException('Name is not empty!');
+    }
+
     return this.prisma.currencies.create({ data });
   }
 
-  findAll() {
+  async findAll() {
     return this.prisma.currencies.findMany({
       select: {
         id: true,
@@ -23,8 +31,8 @@ export class CurrenciesService {
     });
   }
 
-  findOne(id: number) {
-    return this.prisma.currencies.findUnique({
+  async findOne(id: number) {
+    const currencies = await this.prisma.currencies.findUnique({
       where: {
         id: id,
       },
@@ -33,18 +41,44 @@ export class CurrenciesService {
         name: true,
       },
     });
+
+    if (!currencies) {
+      throw new NotFoundException();
+    }
+
+    return currencies;
   }
 
-  update(id: number, updateCurrencyDto: UpdateCurrencyDto) {
+  async update(id: number, updateCurrencyDto: UpdateCurrencyDto) {
+    const currency = await this.prisma.currencies.findUnique({
+      where: { id },
+    });
+
+    if (!currency) {
+      throw new NotFoundException('Currency not found');
+    }
+
+    if (!updateCurrencyDto.name) {
+      throw new BadRequestException('Name is not empty!');
+    }
+
     return this.prisma.currencies.update({
       where: {
         id: id,
       },
-      data: { name: updateCurrencyDto.name },
+      data: { name: updateCurrencyDto.name, updatedAt: new Date() },
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const currency = await this.prisma.currencies.findUnique({
+      where: { id },
+    });
+
+    if (!currency) {
+      throw new NotFoundException('Currency not found');
+    }
+
     return this.prisma.currencies.delete({
       where: { id },
     });
